html/parsing/html5lib/html-tree/template.dat

#data
<body><template>Hello</template>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <template>
|       content
|         "Hello"

#data
<template>Hello</template>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|     <template>
|       content
|         "Hello"
|   <body>

#data
<template></template><div></div>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|     <template>
|       content
|   <body>
|     <div>

#data
<html><template>Hello</template>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|     <template>
|       content
|         "Hello"
|   <body>

#data
<head><template><div></div></template></head>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|     <template>
|       content
|         <div>
|   <body>

#data
<div><template><div><span></template><b>
#errors
no DOCTYPE
no </span>
no </b>
#document
| <html>
|   <head>
|   <body>
|     <div>
|       <template>
|         content
|           <div>
|             <span>
|       <b>

#data
<div><template></div>Hello
#errors
no DOCTYPE
stray </div>
no </template>
no </div>
#document
| <html>
|   <head>
|   <body>
|     <div>
|       <template>
|         content
|           "Hello"

#data
<div></template></div>
#errors
no DOCTYPE
stray </template>
#document
| <html>
|   <head>
|   <body>
|     <div>

#data
<table><template></template></table>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <template>
|         content

#data
<table><template></template></div>
#errors
no DOCTYPE
</div> in table
stray </div>
no </table>
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <template>
|         content

#data
<table><div><template></template></div>
#errors
no DOCTYPE
<div> in table
</div> in table
no </table>
#document
| <html>
|   <head>
|   <body>
|     <div>
|       <template>
|         content
|     <table>

#data
<table><template></template><div></div>
#errors
no DOCTYPE
<div> in table
</div> in table
no </table>
#document
| <html>
|   <head>
|   <body>
|     <div>
|     <table>
|       <template>
|         content

#data
<table>   <template></template></table>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <table>
|       "   "
|       <template>
|         content

#data
<table><tbody><template></template></tbody>
#errors
no DOCTYPE
no </table>
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <template>
|           content

#data
<table><tbody><template></tbody></template>
#errors
no DOCTYPE
stray </tbody>
no </table>
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <template>
|           content

#data
<table><tbody><template></template></tbody></table>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <template>
|           content

#data
<table><thead><template></template></thead>
#errors
no DOCTYPE
no </table>
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <thead>
|         <template>
|           content

#data
<table><tfoot><template></template></tfoot>
#errors
no DOCTYPE
no </table>
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tfoot>
|         <template>
|           content

#data
<select><template></template></select>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <select>
|       <template>
|         content

#data
<select><template><option></option></template></select>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <select>
|       <template>
|         content
|           <option>

#data
<template><option></option></select><option></option></template>
#errors
no DOCTYPE
stray </select>
#document
| <html>
|   <head>
|     <template>
|       content
|         <option>
|         <option>
|   <body>

#data
<select><template></template><option></select>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <select>
|       <template>
|         content
|       <option>

#data
<select><option><template></template></select>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <select>
|       <option>
|         <template>
|           content

#data
<select><template>
#errors
no DOCTYPE
no </template>
no </select>
#document
| <html>
|   <head>
|   <body>
|     <select>
|       <template>
|         content

#data
<select><option></option><template>
#errors
no DOCTYPE
no </template>
no </select>
#document
| <html>
|   <head>
|   <body>
|     <select>
|       <option>
|       <template>
|         content

#data
<select><option></option><template><option>
#errors
no DOCTYPE
no </option>
no </template>
no </select>
#document
| <html>
|   <head>
|   <body>
|     <select>
|       <option>
|       <template>
|         content
|           <option>

#data
<table><thead><template><td></template></table>
#errors
no DOCTYPE
no </td> at </template>
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <thead>
|         <template>
|           content
|             <td>

#data
<table><template><thead></template></table>
#errors
no DOCTYPE
no </thead> at </template>
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <template>
|         content
|           <thead>

#data
<body><table><template><td></tr><div></template></table>
#errors
no DOCTYPE
stray </tr>
no </div>
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <template>
|         content
|           <td>
|             <div>

#data
<table><template><thead></template></thead></table>
#errors
no DOCTYPE
no </thead> at </template>
stray </thead>
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <template>
|         content
|           <thead>

#data
<table><thead><template><tr></template></table>
#errors
no DOCTYPE
no </tr> at </template>
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <thead>
|         <template>
|           content
|             <tr>

#data
<table><template><tr></template></table>
#errors
no DOCTYPE
no </tr>
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <template>
|         content
|           <tr>

#data
<table><tr><template><td>
#errors
no DOCTYPE
no </template> (in body)
no </template>
no </table>
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <template>
|             content
|               <td>

#data
<table><template><tr><template><td></template></tr></template></table>
#errors
no DOCTYPE
no </td>
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <template>
|         content
|           <tr>
|             <template>
|               content
|                 <td>

#data
<table><template><tr><template><td></td></template></tr></template></table>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <template>
|         content
|           <tr>
|             <template>
|               content
|                 <td>

#data
<table><template><td></template>
#errors
no DOCTYPE
no </td>
no </table>
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <template>
|         content
|           <td>

#data
<body><template><td></td></template>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <template>
|       content
|         <td>

#data
<body><template><template><tr></tr></template><td></td></template>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <template>
|       content
|         <template>
|           content
|             <tr>
|         <td>

#data
<table><colgroup><template><col>
#errors
no DOCTYPE
no </template> (in body)
no </template>
no </colgroup>
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <colgroup>
|         <template>
|           content
|             <col>

#data
<frameset><template><frame></frame></template></frameset>
#errors
no DOCTYPE
<template> in frameset
stray </frame>
</template> in frameset
#document
| <html>
|   <head>
|   <frameset>
|     <frame>

#data
<template><frame></frame></frameset><frame></frame></template>
#errors
no DOCTYPE
<frame> in body
stray </frame>
stray </frameset>
<frame> in body
stray </frame>
#document
| <html>
|   <head>
|     <template>
|       content
|   <body>

#data
<template><div><frameset><span></span></div><span></span></template>
#errors
no DOCTYPE
<frameset> in body
#document
| <html>
|   <head>
|     <template>
|       content
|         <div>
|           <span>
|         <span>
|   <body>

#data
<body><template><div><frameset><span></span></div><span></span></template></body>
#errors
no DOCTYPE
<frameset> in body
#document
| <html>
|   <head>
|   <body>
|     <template>
|       content
|         <div>
|           <span>
|         <span>

#data
<body><template><script>var i = 1;</script><td></td></template>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <template>
|       content
|         <script>
|           "var i = 1;"
|         <td>

#data
<body><template><tr><div></div></tr></template>
#errors
no DOCTYPE
<div> in table
</div> in table
#document
| <html>
|   <head>
|   <body>
|     <template>
|       content
|         <tr>
|         <div>

#data
<body><template><tr></tr><td></td></template>
#errors
no DOCTYPE
no <tr>
no </tr>
#document
| <html>
|   <head>
|   <body>
|     <template>
|       content
|         <tr>
|         <tr>
|           <td>

#data
<body><template><td></td></tr><td></td></template>
#errors
no DOCTYPE
stray </tr>
#document
| <html>
|   <head>
|   <body>
|     <template>
|       content
|         <td>
|         <td>

#data
<body><template><td></td><tbody><td></td></template>
#errors
no DOCTYPE
<tbody> in tr
#document
| <html>
|   <head>
|   <body>
|     <template>
|       content
|         <td>
|         <td>

#data
<body><template><td></td><caption></caption><td></td></template>
#errors
no DOCTYPE
<caption> in table body
stray </caption>
#document
| <html>
|   <head>
|   <body>
|     <template>
|       content
|         <td>
|         <td>

#data
<body><template><td></td><colgroup></caption><td></td></template>
#errors
no DOCTYPE
<colgroup> in table body
stray </caption>
#document
| <html>
|   <head>
|   <body>
|     <template>
|       content
|         <td>
|         <td>

#data
<body><template><td></td></table><td></td></template>
#errors
no DOCTYPE
stray </table>
#document
| <html>
|   <head>
|   <body>
|     <template>
|       content
|         <td>
|         <td>

#data
<body><template><tr></tr><tbody><tr></tr></template>
#errors
no DOCTYPE
<tbody> in table body
#document
| <html>
|   <head>
|   <body>
|     <template>
|       content
|         <tr>
|         <tr>

#data
<body><template><tr></tr><caption><tr></tr></template>
#errors
no DOCTYPE
<caption> in table body
#document
| <html>
|   <head>
|   <body>
|     <template>
|       content
|         <tr>
|         <tr>

#data
<body><template><tr></tr></table><tr></tr></template>
#errors
no DOCTYPE
stray </table>
#document
| <html>
|   <head>
|   <body>
|     <template>
|       content
|         <tr>
|         <tr>

#data
<body><template><thead></thead><caption></caption><tbody></tbody></template>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <template>
|       content
|         <thead>
|         <caption>
|         <tbody>

#data
<body><template><thead></thead></table><tbody></tbody></template></body>
#errors
no DOCTYPE
stray </table>
#document
| <html>
|   <head>
|   <body>
|     <template>
|       content
|         <thead>
|         <tbody>

#data
<body><template><div><tr></tr></div></template>
#errors
no DOCTYPE
<tr> in body
stray </tr>
#document
| <html>
|   <head>
|   <body>
|     <template>
|       content
|         <div>

#data
<body><template><em>Hello</em></template>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <template>
|       content
|         <em>
|           "Hello"

#data
<body><template><!--comment--></template>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <template>
|       content
|         <!-- comment -->

#data
<body><template><style></style><td></td></template>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <template>
|       content
|         <style>
|         <td>

#data
<body><template><meta><td></td></template>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <template>
|       content
|         <meta>
|         <td>

#data
<body><template><link><td></td></template>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <template>
|       content
|         <link>
|         <td>

#data
<body><template><template><tr></tr></template><td></td></template>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <template>
|       content
|         <template>
|           content
|             <tr>
|         <td>

#data
<body><table><colgroup><template><col></col></template></colgroup></table></body>
#errors
no DOCTYPE
stray </col>
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <colgroup>
|         <template>
|           content
|             <col>

#data
<body a=b><template><div></div><body c=d><div></div></body></template></body>
#errors
no DOCTYPE
<body>
</body>
#document
| <html>
|   <head>
|   <body>
|     a="b"
|     <template>
|       content
|         <div>
|         <div>

#data
<html a=b><template><div><html b=c><span></template>
#errors
no DOCTYPE
<html>
no </span>
#document
| <html>
|   a="b"
|   <head>
|     <template>
|       content
|         <div>
|           <span>
|   <body>

#data
<html a=b><template><col></col><html b=c><col></col></template>
#errors
no DOCTYPE
stray </col>
<html>
</col>
#document
| <html>
|   a="b"
|   <head>
|     <template>
|       content
|         <col>
|         <col>
|   <body>

#data
<html a=b><template><frame></frame><html b=c><frame></frame></template>
#errors
no DOCTYPE
<frame> in body
stray </frame>
<html> ignored
<frame> in body
stray </frame>
#document
| <html>
|   a="b"
|   <head>
|     <template>
|       content
|   <body>

#data
<body><template><tr></tr><template></template><td></td></template>
#errors
no DOCTYPE
no <tr>
no </tr> before </template>
#document
| <html>
|   <head>
|   <body>
|     <template>
|       content
|         <tr>
|         <template>
|           content
|         <tr>
|           <td>

#data
<body><template><thead></thead><template><tr></tr></template><tr></tr><tfoot></tfoot></template>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <template>
|       content
|         <thead>
|         <template>
|           content
|             <tr>
|         <tbody>
|           <tr>
|         <tfoot>

#data
<body><template><template><b><template></template></template>text</template>
#errors
no DOCTYPE
no </b> at </template>
#document
| <html>
|   <head>
|   <body>
|     <template>
|       content
|         <template>
|           content
|             <b>
|               <template>
|                 content
|         "text"

#data
<body><template><col><colgroup>
#errors
no DOCTYPE
<colgroup>
no </template> (in body)
no </template>
#document
| <html>
|   <head>
|   <body>
|     <template>
|       content
|         <col>

#data
<body><template><col></colgroup>
#errors
no DOCTYPE
</colgroup>
no </template> (in body)
no </template>
#document
| <html>
|   <head>
|   <body>
|     <template>
|       content
|         <col>

#data
<body><template><col><colgroup></template></body>
#errors
no DOCTYPE
<colgroup>
#document
| <html>
|   <head>
|   <body>
|     <template>
|       content
|         <col>

#data
<body><template><col><div>
#errors
no DOCTYPE
<div> in colgroup
no </div>
no </template>
#document
| <html>
|   <head>
|   <body>
|     <template>
|       content
|         <col>

#data
<body><template><col></div>
#errors
no DOCTYPE
stray </div>
no </template> (in body)
no </template>
#document
| <html>
|   <head>
|   <body>
|     <template>
|       content
|         <col>

#data
<body><template><col>Hello
#errors
no DOCTYPE
"Hello" in colgroup
no </template> (in body)
no </template>
#document
| <html>
|   <head>
|   <body>
|     <template>
|       content
|         <col>

#data
<body><template><i><menu>Foo</i>
#errors
no DOCTYPE
AAA at </i>
no </menu>
no </template>
#document
| <html>
|   <head>
|   <body>
|     <template>
|       content
|         <i>
|         <menu>
|           <i>
|             "Foo"

#data
<body><template></div><div>Foo</div><template></template><tr></tr>
#errors
no DOCTYPE
stray </div>
<tr>
stray </tr>
no </template> (in body)
no </template>
#document
| <html>
|   <head>
|   <body>
|     <template>
|       content
|         <div>
|           "Foo"
|         <template>
|           content

#data
<body><div><template></div><tr><td>Foo</td></tr></template>
#errors
no DOCTYPE
stray </div>
no </div>
#document
| <html>
|   <head>
|   <body>
|     <div>
|       <template>
|         content
|           <tr>
|             <td>
|               "Foo"

#data
<template></figcaption><sub><table></table>
#errors
no DOCTYPE
stray </figcaption>
no </sub>
no </template>
#document
| <html>
|   <head>
|     <template>
|       content
|         <sub>
|           <table>
|   <body>

#data
<template><template>
#errors
no DOCTYPE
no </template>
no </template>
#document
| <html>
|   <head>
|     <template>
|       content
|         <template>
|           content
|   <body>

#data
<template><div>
#errors
no DOCTYPE
no </div>
no </template>
#document
| <html>
|   <head>
|     <template>
|       content
|         <div>
|   <body>

#data
<template><template><div>
#errors
no DOCTYPE
no </div>
no </template>
no </template>
#document
| <html>
|   <head>
|     <template>
|       content
|         <template>
|           content
|             <div>
|   <body>

#data
<template><template><table>
#errors
no DOCTYPE
no </table>
no </template>
no </template>
#document
| <html>
|   <head>
|     <template>
|       content
|         <template>
|           content
|             <table>
|   <body>

#data
<template><template><tbody>
#errors
no DOCTYPE
no </template> (in body)
no </template>
no </template>
#document
| <html>
|   <head>
|     <template>
|       content
|         <template>
|           content
|             <tbody>
|   <body>

#data
<template><template><tr>
#errors
no DOCTYPE
no </template> (in body)
no </template>
no </template>
#document
| <html>
|   <head>
|     <template>
|       content
|         <template>
|           content
|             <tr>
|   <body>

#data
<template><template><td>
#errors
no DOCTYPE
no </template> (in body)
no </template>
no </template>
#document
| <html>
|   <head>
|     <template>
|       content
|         <template>
|           content
|             <td>
|   <body>

#data
<template><template><caption>
#errors
no DOCTYPE
no </caption>
no </template>
no </template>
#document
| <html>
|   <head>
|     <template>
|       content
|         <template>
|           content
|             <caption>
|   <body>

#data
<template><template><colgroup>
#errors
no DOCTYPE
no </colgroup>
no </template>
no </template>
#document
| <html>
|   <head>
|     <template>
|       content
|         <template>
|           content
|             <colgroup>
|   <body>

#data
<template><template><col>
#errors
no DOCTYPE
no </template> (in body)
no </template>
no </template>
#document
| <html>
|   <head>
|     <template>
|       content
|         <template>
|           content
|             <col>
|   <body>

#data
<template><template><tbody><select>
#errors
no DOCTYPE
<select> in table
no </select> at EOF
no </template>
no </template>
#document
| <html>
|   <head>
|     <template>
|       content
|         <template>
|           content
|             <tbody>
|             <select>
|   <body>

#data
<template><template><table>Foo
#errors
no DOCTYPE
"Foo" in table
no </table>
no </template>
no </template>
#document
| <html>
|   <head>
|     <template>
|       content
|         <template>
|           content
|             "Foo"
|             <table>
|   <body>

#data
<template><template><frame>
#errors
no DOCTYPE
<frame> in body
no </template> (in body)
no </template>
no </template>
#document
| <html>
|   <head>
|     <template>
|       content
|         <template>
|           content
|   <body>

#data
<template><template><script>var i
#errors
no DOCTYPE
no </script> at EOF
no </template> at EOF
no </template> at EOF
#document
| <html>
|   <head>
|     <template>
|       content
|         <template>
|           content
|             <script>
|               "var i"
|   <body>

#data
<template><template><style>var i
#errors
no DOCTYPE
no </style> at EOF
no </template> at EOF
no </template> at EOF
#document
| <html>
|   <head>
|     <template>
|       content
|         <template>
|           content
|             <style>
|               "var i"
|   <body>

#data
<template><table></template><body><span>Foo
#errors
no DOCTYPE
no </table> at </template>
no </span> at EOF
#document
| <html>
|   <head>
|     <template>
|       content
|         <table>
|   <body>
|     <span>
|       "Foo"

#data
<template><td></template><body><span>Foo
#errors
no DOCTYPE
no </td> at </template>
no </span> at EOF
#document
| <html>
|   <head>
|     <template>
|       content
|         <td>
|   <body>
|     <span>
|       "Foo"

#data
<template><object></template><body><span>Foo
#errors
no DOCTYPE
not closed before ancestor end tag:object at </template>
no </span> at EOF
#document
| <html>
|   <head>
|     <template>
|       content
|         <object>
|   <body>
|     <span>
|       "Foo"

#data
<template><svg><template>
#errors
no DOCTYPE
no end tag at EOF <svg template>
no end tag at EOF <template>
#document
| <html>
|   <head>
|     <template>
|       content
|         <svg svg>
|           <svg template>
|   <body>

#data
<template><svg><foo><template><foreignObject><div></template><div>
#errors
no DOCTYPE
not closed before ancestor end tag:div at </template>
no end tag before EOF:div
#document
| <html>
|   <head>
|     <template>
|       content
|         <svg svg>
|           <svg foo>
|             <svg template>
|               <svg foreignObject>
|                 <div>
|   <body>
|     <div>

#data
<dummy><template><span></dummy>
#errors
no DOCTYPE
stray end tag </dummy>
no </span> at EOF
no </template> at EOF
no </dummy> at EOF
#document
| <html>
|   <head>
|   <body>
|     <dummy>
|       <template>
|         content
|           <span>

#data
<body><table><tr><td><select><template>Foo</template><caption>A</table>
#errors
no DOCTYPE
(1,62): unexpected-caption-in-select-in-table
(1,71): unexpected-table-end-in-caption
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <select>
|               <template>
|                 content
|                   "Foo"
|       <caption>
|         "A"

#data
<body></body><template>
#errors
no DOCTYPE
(1,23): template-after-body
(1,24): eof-in-template
#document
| <html>
|   <head>
|   <body>
|     <template>
|       content

#data
<head></head><template>
#errors
no DOCTYPE
(1,23): template-after-head
(1,24): eof-in-template
#document
| <html>
|   <head>
|     <template>
|       content
|   <body>

#data
<head></head><template>Foo</template>
#errors
no DOCTYPE
(1,23): template-after-head
#document
| <html>
|   <head>
|     <template>
|       content
|         "Foo"
|   <body>
